Use artifacts volume for logs, test results, etc. (#1)
Assumes applications are writing build artifacts to an `artifacts:/opt/app/artifacts` volume, copying that out of application containers and storing it for reference. This makes it possible for applications to distinguish different test outputs, which is not possible when piping everything to stdout/err.

Applications are required to implement this volume in their docker-compose.ci.yml files.